backend/requirements.txt
Django==4.0
djangorestframework>=3.12.4
djangorestframework-simplejwt==5.3.1
gunicorn>=20.0.4
python-dotenv>=0.19.2
django-cors-headers==4.3.1
google-cloud-ndb==2.3.1
bcrypt==4.1.3
